TreeFormer/requirements.txt
numpy==2.2.6
Pillow==9.4.0
scikit_learn==1.2.2
scipy==1.14.1
timm==0.4.12
torch==1.12.1
torchvision==0.13.1
h5py==3.8.0
matplotlib==3.10.1
